Cross Shot Changes
Hopefully it won't explode

+++ src/Weapons/Buster.cs
@@ -90,7 +90,7 @@ public class Buster : Weapon {
 				0 => "busterX3",
 				1 => "buster2X3",
 				2 => "buster3X3",
-				3 => "buster3X3",
+				3 => "",
 				_ => shootSound
 			};
 		}
@@ -171,7 +171,6 @@ public class Buster : Weapon {
 			player.character.playSound(shootSound, sendRpc: true);
 		}
 	}
-	
 	
 	public void createBuster4Line(
 		float x, float y, int xDir, Player player,
@@ -277,8 +276,11 @@ public class Buster3Proj : Projectile {
 	public List<Sprite> spriteMids = new List<Sprite>();
 	float partTime;
 
-	public Buster3Proj(Weapon weapon, Point pos, int xDir, int type, Player player, ushort netProjId, bool rpc = false) :
-		base(weapon, pos, xDir, 350, 3, player, "buster3", Global.defFlinch, 0f, netProjId, player.ownedByLocalPlayer) {
+	public Buster3Proj(
+		Weapon weapon, Point pos, int xDir, int type, Player player, ushort netProjId, bool rpc = false
+	) : base(
+		weapon, pos, xDir, 350, 3, player, "buster3", Global.defFlinch, 0f, netProjId, player.ownedByLocalPlayer
+	) {
 		this.type = type;
 		maxTime = 0.5f;
 		fadeSprite = "buster3_fade";
@@ -293,11 +295,6 @@ public class Buster3Proj : Projectile {
 		// Regular yellow charge
 		if (type == 0) {
 			damager.flinch = Global.halfFlinch;
-			damager.damage = 3;
-			projId = (int)ProjIds.Buster3;
-			reflectable = true;
-			fadeSprite = "buster3_fade";
-			fadeOnAutoDestroy = true;
 			if (player.hasArmArmor(ArmorId.Giga)) {
 				changeSprite("buster3_x2", true);
 				}
@@ -305,8 +302,10 @@ public class Buster3Proj : Projectile {
 				changeSprite("buster3_x3", true);
 			}
 		}
+		
 		// Double buster part 1
 		if (type == 1) {
+			damager.flinch = Global.defFlinch;
 			damager.damage = 4;
 			changeSprite("buster3_x2", true);
 			projId = (int)ProjIds.Buster4;
@@ -331,6 +330,7 @@ public class Buster3Proj : Projectile {
 			changeSprite("buster4_x3", true);
 			fadeSprite = "buster4_x2_fade";
 			vel.x = 0;
+			maxTime = 1f;
 			projId = (int)ProjIds.Buster4;
 			reflectable = false;
 			fadeOnAutoDestroy = true;
@@ -359,20 +359,30 @@ public class Buster3Proj : Projectile {
 	// Down here is where the Cross Shot actually happens
 	public override void onCollision(CollideData other) {
 		base.onCollision(other);	
-				if (type == 4) {
+		
 				if (!ownedByLocalPlayer) return;
-				if (other.gameObject is BusterX3Proj1 && ownedByLocalPlayer && !destroyed) {
-					fadeSprite = "buster4_x3_muzzle"; destroySelf();
+				if (other.gameObject is BusterX3Proj1 X3shot && X3shot.ownedByLocalPlayer && !destroyed) {
+					Global.level.delayedActions.Add(new DelayedAction(delegate {
+					new Anim(new Point(pos.x, pos.y), "buster4_x3_muzzle", xDir, null, true);
+					fadeSprite = null; X3shot.fadeSprite = null;
+					destroySelf(); X3shot.destroySelf();
 					Global.level.delayedActions.Add(new DelayedAction(delegate { 
+					if (!owner.hasUltimateArmor()) {
 					new Buster3Proj(weapon, pos, xDir, 5, owner, owner.getNextActorNetId(), rpc: true);
+					} else {
+					// new Anim(pos, "buster4_muzzle_flash", xDir, null, true);
+					new BusterPlasmaProj(weapon, pos, xDir, owner, owner.getNextActorNetId(), rpc: true);
+					playSound("plasmaShot", sendRpc: true);
+					}
 					new BusterX3Proj3(weapon, pos, xDir, 0, owner, owner.getNextActorNetId(), rpc: true);
 					new BusterX3Proj3(weapon, pos, xDir, 1, owner, owner.getNextActorNetId(), rpc: true);
 					new BusterX3Proj3(weapon, pos, xDir, 2, owner, owner.getNextActorNetId(), rpc: true);
 					new BusterX3Proj3(weapon, pos, xDir, 3, owner, owner.getNextActorNetId(), rpc: true);
 					}, 20f / 60f ));
+					}, 1f / 60f ));
 			}
 	}
-}
+
 	public override void update() {
 		base.update();
 		if (type == 3) {
@@ -603,27 +613,19 @@ public class X3ChargeShot : CharState {
 		if (!fired && character.currentFrame.getBusterOffset() != null && player.ownedByLocalPlayer) {
 			fired = true;
 			if (state == 0) {
-				Point shootPos = character.getShootPos();
-				int shootDir = character.getShootXDir();
-				if (!player.hasUltimateArmor()) {
-					new BusterX3Proj1(player.weapon, shootPos, shootDir, 0, player, player.getNextActorNetId(), rpc: true);
-					character.playSound("buster3X3", sendRpc: true);
-				} else {
-					new Anim(shootPos, "buster4_muzzle_flash", shootDir, null, true);
-					new BusterPlasmaProj(
-						player.weapon, shootPos, shootDir,
-						player, player.getNextActorNetId(), rpc: true
-					);
-					character.playSound("plasmaShot", sendRpc: true);
-				}
+				new BusterX3Proj1(
+					player.weapon, character.getShootPos(), character.getShootXDir(), 0,
+					player, player.getNextActorNetId(), rpc: true);
+				character.playSound("buster3X3", sendRpc: true);
+				
 			} else {
 				if (hyperBusterWeapon != null) {
 					hyperBusterWeapon.ammo -= hyperBusterWeapon.getChipFactoredAmmoUsage(player);
 				}
-				float xDir = character.getShootXDir();
+				// float xDir = character.getShootXDir();
 
 				new Buster3Proj(
-					player.weapon, character.getShootPos(), character.getShootXDir(), 4,
+					player.weapon, character.getShootPos(), character.getShootXDir(), 0,
 					player, player.getNextActorNetId(), rpc: true
 				);
 				character.playSound("buster3X3", sendRpc: true);
@@ -722,7 +724,7 @@ public class BusterX3Proj1 : Projectile {
 	float line3Y = 2;
 	float partTime;
 		public BusterX3Proj1(Weapon weapon, Point pos, int xDir, int type, Player player, ushort netProjId, bool rpc = false) :
-		base(weapon, pos, xDir, 350, 1, player, "buster2", Global.halfFlinch, 0f, netProjId, player.ownedByLocalPlayer) {
+		base(weapon, pos, xDir, 350, 1, player, "buster4_max_orb2", Global.halfFlinch, 0f, netProjId, player.ownedByLocalPlayer) {
 		this.type = type;
 		maxTime = 1.05f;
 		vel.x = 0;
@@ -730,12 +732,7 @@ public class BusterX3Proj1 : Projectile {
 		fadeOnAutoDestroy = true;
 		projId = (int)ProjIds.BusterX3Proj1;
 		reflectable = false;
-		// probably could delete this part as well
-		for (int i = 0; i < 6; i++) {
-				var midSprite = Global.sprites["buster4_max_orb2"].clone();
-				spriteMids.Add(midSprite);
-			}
-
+		
 		if (rpc) {
 			rpcCreate(pos, player, netProjId, xDir, (byte)type);
 		}
@@ -761,55 +758,25 @@ public class BusterX3Proj1 : Projectile {
 		new Anim(new Point(pos.x, pos.y - 4 + (line2Y * -1)), "buster4_max_orb2", xDir, null, true, zIndex == zLayer);
 		new Anim(new Point(pos.x + 4, pos.y + line3Y), "buster4_max_orb3", xDir, null, true, zIndex == zLayer);
 	}
-	// might also have to delete this
-	 /* public override void render(float x, float y) {
-		base.render(x, y);
-			float piHalf = MathF.PI / 2;
-			float xOffset = 8;
-			float partTime = (time * 0.75f);
-			for (int i = 0; i < 6; i++) {
-				float t = 0;
-				float xOff2 = 0;
-				float sinX = 0;
-				if (i < 3) {
-					t = partTime - (i * 0.025f);
-					xOff2 = i * xDir * 3;
-					sinX = 5 * MathF.Cos(partTime * 20);
-				} else {
-					t = partTime + (MathF.PI / 4) - ((i - 3) * 0.025f);
-					xOff2 = (i - 3) * xDir * 3;
-					sinX = 5 * MathF.Sin((partTime) * 20);
-				}
-				float sinY = 15 * MathF.Sin(t * 20);
-				long zIndexTarget = zIndex - 1;
-				float currentOffset = (t * 20) % (MathF.PI * 2);
-				if (currentOffset > piHalf && currentOffset < piHalf * 3) {
-					zIndexTarget = zIndex + 1;
-				}
-				spriteMids[i].draw(
-					spriteMids[i].frameIndex,
-					pos.x + x + sinX - xOff2 + xOffset,
-					pos.y + y + sinY, xDir, yDir,
-					getRenderEffectSet(), 1, 1, 1, zIndexTarget
-				);
-				spriteMids[i].update();
-			}
-		} */
+	/*
 	public override void onCollision(CollideData other) {
 		base.onCollision(other);
-		if(ownedByLocalPlayer && !destroyed)
-			if(other.gameObject is Buster3Proj){
-				Global.level.delayedActions.Add(new DelayedAction(delegate { 
-					fadeSprite = null;
-					// is there any better code to use no fadeSprite?
-					destroySelf();
-					}, 2f / 60f ));
-			}
+		if(other.gameObject is Buster3Proj X3shot && X3shot.ownedByLocalPlayer && !destroyed){
+			Global.level.delayedActions.Add(new DelayedAction(delegate { 
+			// fadeSprite = null;
+			// is there any better code to use no fadeSprite?
+			destroySelf(); 
+			// X3shot.destroySelf();
+			}, 1f / 60f ));
 		}
+	}
+	*/
+
 	// This down here is meant for the split shot when hitting another character
 	public override void onHitDamagable(IDamagable damagable) {
 		base.onHitDamagable(damagable);
 		if (ownedByLocalPlayer) {
+				fadeSprite = "buster3_fade";
 				destroySelf();
 				Global.level.delayedActions.Add(new DelayedAction(delegate { 
 					new BusterX3Proj2(weapon, pos, xDir, 0, owner, owner.getNextActorNetId(), rpc: true);
@@ -826,7 +793,7 @@ public class BusterX3Proj2 : Projectile {
 		Weapon weapon, Point pos, int xDir, int type,
 		Player player, ushort netProjId, bool rpc = false
 	) : base(
-		weapon, pos, xDir, 400, 2,
+		weapon, pos, xDir, 400, 1,
 		player, type == 0 || type == 1 ? "buster4_max_orb1" : "buster4_max_orb3",
 		0, 0, netProjId, player.ownedByLocalPlayer
 	) {
